Add support for achievements
Manage all achievements with new AchievementTracker class, which gets notifications from
LevelLayer and GameFlow.

Store state of achievements in persistent store to keep track of them even if game is exited.

Add 'About layer', in the future it will contain credits, but for now it has only a debug button
(with settings-like icon) resetting the state of the game. Currently all achievements are erased
and IAP is removed.

'Pro Runner' achievement is still TODO.