New animations:
There's a new tutorial00.x animation, which shows waving rat, intended to be
used on a tutorial's welcome screen. As we already had tutorial00.x animation,
all numbers of existing tutorial animations were shifted by 1, i.e.  animation
explaining vial, which used to be named tutorial01.x is now called
tutorial02.x.

In addition, score events were improved in a tutorial01.x animation, so the
accessory score display in a tutorial level should now behave in a better way.

Two new animtaions showing happpy rat were added, intended to be used on a post
level layer after the user has beaten  own record or won a challenge:
post_happy_01.x and post_happy_02.x. We should randomly select one of them in
the post level layer. Also, one of them (probably post_happy_02.x) can be used
as a last animation in the tutorial.

The same story with the sad rat, which has failed to beat the record or has
lost the challenge: post_sad_01.x and post_sad_02.x/post_sad_03.x. The latter
two are almost the same, but a creative director and a animator could not come
to conclusion which one is better, so it's up to you to select one of them,
dear programmer.